+++ beakpeek/lib/View/Home/home.dart
@@ -1,6 +1,6 @@
 import 'package:beakpeek/Styles/home_page_styles.dart';
 import 'package:flutter/material.dart';
-import 'package:beakpeek/LandingText/cust_buttons.dart';
+import 'package:beakpeek/LandingText/custom_buttons.dart';
 import 'package:beakpeek/LandingText/nav.dart';
 
 class Home extends StatelessWidget {
